+++ lib/provider.js
@@ -1,7 +1,8 @@
 'use babel';
 
 import {filter} from 'fuzzaldrin';
-import {Range} from 'atom';
+import {dispatch} from './functional';
+import cmake_context from './cmake-context';
 import commands from './commands';
 import constants from './constants';
 import find_modules from './find-modules';
@@ -12,24 +13,8 @@ export const selector = '.source.cmake';
 export const disableForSelector = '.source.cmake .comment';
 export const inclusionPriority = 1;
 
-function existy(value) {
-    return value != null;
-}
-
-function dispatch() {
-    var funs = arguments;
-    return function() {
-        for (var f of funs) {
-            var ret = f.apply(null, arguments);
-            if (existy(ret)) {
-                return ret;
-            }
-        }
-    };
-}
-
 function find_modules_suggestions(prefix) {
-    return filter(find_modules, prefix.toLowerCase(), {key: 'id'}).map((module) => ({
+    return filter(find_modules, prefix).map((module) => ({
         text: module.name,
         displayText: module.name,
         type: 'constant'
@@ -68,56 +53,30 @@ function commands_suggestions(prefix) {
     }));
 }
 
-function skip(i, predicate) {
-    while (i >= 0 && predicate(i)) {
-        i--;
-    }
-    return i;
-}
-
-function function_scope_name(editor) {
-    const range = new Range([0, 0], editor.getCursorBufferPosition());
-    const text = editor.getTextInRange(range);
-    const last = skip(
-        skip(text.length - 1, index => text.charAt(index) !== '(') - 1,
-        index => /\s/.test(text.charAt(index))
-    ) + 1;
-    const first = skip(
-        last - 1,
-        index => !/\s/.test(text.charAt(index))
-    );
-    return text.substring(first, last).trim();
-}
-
-const suggest_function_scope = dispatch(
-    (name, prefix) => {
-        if (name === 'find_package') {
+const suggest = dispatch(
+    (context, prefix) => {
+        if (context === 'find_package') {
             return find_modules_suggestions(prefix);
         }
     },
-    (name, prefix) => {
-        if (name === 'include') {
+    (context, prefix) => {
+        if (context === 'include') {
             return modules_suggestions(prefix);
         }
     },
-    (name, prefix) => [
-        ...constants_suggestions(prefix),
-        ...variables_suggestions(prefix)
-    ]
-);
-
-const suggest = dispatch(
-    (editor, prefix, scope_descriptor) => {
-        if (scope_descriptor.scopes.length > 1) {
-            const scope_name = function_scope_name(editor, prefix);
-            return suggest_function_scope(scope_name, prefix);
+    (context, prefix) => {
+        if (context.length > 0) {
+            return [
+                ...constants_suggestions(prefix),
+                ...variables_suggestions(prefix)
+            ];
         }
     },
-    (editor, prefix) => commands_suggestions(prefix)
+    (context, prefix) => commands_suggestions(prefix)
 )
 
 export function getSuggestions({editor, prefix, scopeDescriptor: scope_descriptor}) {
-    return suggest(editor, prefix, scope_descriptor);
+    return suggest(cmake_context(editor, scope_descriptor), prefix);
 }
 
 export function onDidInsertSuggestion({editor, suggestion}) {
